dom/indexedDB/test/test_null_keys.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const description = "My Test Database";

      let request = moz_indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let objectStore = db.createObjectStore("foo", "");

      request = objectStore.add({}, null);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      ok(event.result != null, "Valid key");

      let key1 = event.result;

      request = objectStore.add({}, null);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      ok(event.result != null, "Valid key");

      ok(key1 != event.result, "Different unique keys");

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
